frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["ssgrummons <ssgrummo@us.ibm.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
streamlit = "^1.44.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.0"
pytest-cov = "^4.1.0"  # Optional, for coverage reporting

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
